Old code relied on sending messages to get certain goroutines to quit, but there were multiple issues.
The channel used could be read by multiple goroutines at the same time, resulting in the 'special' goroutine not getting the message and never exiting.
This would result in the code attempting to exit but would subsequently send a message back to the goroutine that was supposed to be exiting.

This fixes the potential deadlock, leak, and non-deterministic channel read. It also makes other channel sends safe.