This change partially reverts 22b3887f319900712cd2285bebb16d979e1a324c
While the change did fix the problem, it wasn't the best solution. The solution to making the locks exclusive
was not to spawn a new goroutine but to unlock the first lock after the data it was guarding is dealt with instead
of unlocking in a defer. Additionally what lock contention there was should have been dealt with in a newer but
previous commit to the one mentioned.